mark_agent/pyproject.toml
[project]
name = "mark-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.24",
    "crewai[google-genai]>=1.10.0",
    "python-dotenv>=1.1.1",
]
